tests: add asserts to kinetic_sdk test suite
